TissueGrowth.jl

TissueGrowth.jl is a open source project package developed to simulate the evoluition of biological tissue interface during tissue growth. Our model considers mechanical interactions between neighbouring cells and a secretion rate of new tissue material that is proportional to cell density. In this project we include cell proliferation, apoptosis (death) and embedment as stochastic processes. Below is a discription of our model.

Model Description

This cell-based mathematical model

Calculates the mechanical relaxation and normal velocity in a system with periodic boundary conditions. The derivatives are based on the discrete equation given by:

\[\frac{\text{d}\mathbf{u}ᵢ}{\text{d}t} = \frac{1}{η}\bigg((\mathbf{F}ₛ⁺ - \mathbf{F}ₛ⁻)\cdot\bm{τ}\bigg)\bm{τ} + Vₙ\mathbf{n} \]

where,

\[\mathbf{F}ₛ⁺ = f(\lVert \mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ \rVert) \frac{\mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ}{\lVert \mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ \rVert}, \hspace{0.5cm} \mathbf{F}ₛ⁻ = f(\lVert \mathbf{u}ᵢ - \mathbf{u}ᵢ₋₁ \rVert) \frac{\mathbf{u}ᵢ - \mathbf{u}ᵢ₋₁}{\lVert \mathbf{u}ᵢ - \mathbf{u}ᵢ₋₁ \rVert}\]

In this current version of the simulation code we use a nonlinear resorting force given by,

\[f(\lVert \mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ \rVert) = kₛ \bigg(\frac{1}{l_{0}} - \frac{1}{\lVert \mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ \rVert }\bigg)\]

and a normal velocity which is proportional to density such that,

\[Vₙ = k_{f}\lVert \mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ \rVert\]

given $f(\lVert \mathbf{u}ᵢ₊₁ - \mathbf{u}ᵢ \rVert)$ is the restoration force function and $Vₙ$ is the velocity in the normal direction.

Documentation Index

TissueGrowth.SimResults_t

Base.deleteat!

Base.insert!

TissueGrowth.A

TissueGrowth.E

TissueGrowth.Fₛ⁺

TissueGrowth.Fₛ⁻

TissueGrowth.ODE_fnc_1D!

TissueGrowth.ODE_fnc_2D!

TissueGrowth.P

TissueGrowth.SetupODEproblem1D

TissueGrowth.SetupODEproblem2D

TissueGrowth.Vₙ

TissueGrowth.affect!

TissueGrowth.n

TissueGrowth.sim1D

TissueGrowth.sim2D

TissueGrowth.δ

TissueGrowth.κ

TissueGrowth.ρ

TissueGrowth.τ

Data Structure

TissueGrowth.SimResults_t — Type

SimResults_t

A data structure representing the results of a simulation.

This structure contains various fields that store the results and relevant data from a simulation process. It includes the type of simulation, time steps, state vectors, forces, densities, and other relevant quantities.

Fields

btype: A String indicating the type of simulation or boundary condition.

t: A Vector{Float64} representing the time steps at which the simulation results are recorded.

u: A Vector of ElasticMatrix{Float64,Vector{Float64}} representing the state of the system at each time step.

∑F: A Vector of ElasticVector{Float64,Vector{Float64}}, representing the sum of forces at each time step.

Density: A Vector of ElasticMatrix{Float64,Vector{Float64}} representing the density of the system at each time step.

Vₙ: A Vector{Vector{Float64}} representing the normal velocity at each time step.

Ω: A Vector{Float64} representing the void area within the pore at each time step.

ψ: A Vector of ElasticMatrix{Float64,Vector{Float64}}, representing the stress of the springs/ cells at each time step.

Κ: A Vector of ElasticMatrix{Float64,Vector{Float64}}, representing the approximate curvature of the moving boundary at each time step.

1D simulation code:

TissueGrowth.SetupODEproblem1D — Method

SetupODEproblem1D(btype, M, m, R₀, kₛ, η, kf, l₀, δt, Tmax, growth_dir, dist_type)

Create and configure a 1D ODE problem for mechanical relaxation simulations.

This function sets up the initial conditions and parameters for a 1D ODE problem based on the specified boundary type and other physical parameters.

Arguments

btype: Boundary type (e.g., 'circle', 'triangle').

M, m: Number of springs/particles in the system.

R₀: Initial radius for circular boundary problems.

kₛ, η: Prescaled mechanical relaxation coefficients.

kf: Tissue production rate per cell per unit time.

l₀: Resting spring length.

δt: Euler timestep size.

Tmax: End of simulation time.

growth_dir: Direction of tissue growth ('inward' or 'outward').

dist_type: Distribution type for node placement.

Returns

Configured ODEProblem instance and parameters for the 1D simulation.

TissueGrowth.sim1D — Method

sim1D()

Execute a series of 1D mechanical relaxation simulations.

This function sets up and runs a series of 1D simulations for different stiffness values. It iterates over an array of stiffness coefficients, sets up the corresponding ODE problem for each case, solves it, and collects the results.

Simulation Parameters

N: Number of cells in the simulation.

m: Number of springs per cell.

M: Total number of springs along the interface.

R₀: Radius or characteristic length of the initial shape.

D: Array of diffuision coefficients used to calculate cell stiffness.

l₀: Resting length of the spring per cell.

kf: Tissue production rate per cell.

η: Viscosity or damping coefficient per cell.

Tmax: Total simulation time (in days).

δt: Time step for the Euler integration.

growth_dir: Growth direction for the simulation ('1D').

btype: Type of boundary condition ('SineWave').

savetimes: Time points at which to save the simulation results.

Returns

A vector of SimResults_t objects, each representing the simulation results for a different stiffness coefficient.

2D simulation code:

TissueGrowth.SetupODEproblem2D — Method

SetupODEproblem2D(btype, M, m, R₀, kₛ, η, kf, l₀, δt, Tmax, growth_dir, prolif, death, embed, α, β, γ, dist_type)

Set up and configure a 2D ODE problem for mechanical relaxation simulations in tissue growth.

This function initializes the conditions and parameters for a 2D ODE problem based on the specified boundary type, physical parameters, and cell behaviors. It then constructs an ODEProblem object, ready for solving with DifferentialEquations.jl.

Arguments

btype: Boundary type (e.g., 'circle', 'triangle').

M: Total number of springs along the interface.

m: Number of springs per cell.

R₀: Initial radius or characteristic length of the shape.

kₛ: Spring stiffness coefficient.

η: Viscosity or damping coefficient.

kf: Tissue production rate per cell.

l₀: Resting length of the spring per cell.

δt: Time step for the numerical integration.

Tmax: Total simulation time.

growth_dir: Direction of tissue growth ('inward' or 'outward').

prolif: Boolean flag for cell proliferation.

death: Boolean flag for cell death.

embed: Boolean flag for cell embedding.

α, β, γ: Parameters for the cell behaviors.

dist_type: Distribution type for node placement.

Returns

ODEProblem: An ODE problem instance set up with the specified parameters and initial conditions.

p: A tuple containing the parameters used in setting up the ODE problem.

TissueGrowth.sim2D — Method

sim2D()

Execute a series of 2D mechanical relaxation simulations.

This function sets up and runs 2D simulations for different stiffness coefficients and boundary types. It iterates over arrays of stiffness coefficients and boundary types, sets up the corresponding ODE problem for each case, solves it using a specified numerical method with periodic callbacks, and collects the results.

Simulation Parameters

N: Number of cells in the simulation.

m: Number of springs per cell.

M: Total number of springs along the interface.

R₀: Radius or characteristic length of the initial shape.

D: Array of diffuision coefficients used to calculate cell stiffness.

l₀: Resting length of the spring per cell.

kf: Tissue production rate per cell.

η: Viscosity or damping coefficient per cell.

growth_dir: Direction of tissue growth ('inward' or 'outward').

Tmax: Total simulation time (in days).

δt: Time step for the numerical integration.

btypes: Types of boundary conditions (e.g., 'circle', 'triangle').

dist_type: Distribution type for node placement (e.g., 'Linear', 'sigmoid').

prolif, death, embed: Boolean flags indicating cell behaviors.

α, β, γ: Parameters for cell behaviors.

event_δt: Time interval for periodic callback events.

savetimes: Time points at which to save the simulation results.

Returns

A vector of vectors of SimResults_t objects. Each inner vector represents the simulation results for different boundary types under a specific stiffness coefficient.

Example

all_results = sim2D()

ODE Problem:

TissueGrowth.ODE_fnc_1D! — Method

ODE_fnc_1D!(du, u, p, t)

Define the ODE system for 1D mechanical relaxation with initial conditions. This function computes the derivatives du based on the current state u and parameters p. In this system there is a periodic boundary condition such that x ∈ [0,2π] with i spring boundary nodes and u₋₁ = uₙ, uₙ₊₁ = u₁

Arguments

du: Array to store the derivatives of u.

u: Current state array.

p: Parameters tuple (N, kₛ, η, kf, l₀, δt, growth_dir).

t: Current time.

Description

Calculates the mechanical relaxation and normal velocity in a 1D system with periodic boundary conditions. The derivatives are based on spring forces and mechanical properties defined in p.

TissueGrowth.ODE_fnc_2D! — Method

ODE_fnc_2D!(du, u, p, t)

Define the ODE system for 2D mechanical relaxation with initial conditions. This function computes the derivatives du based on the current state u and parameters p.

Arguments

du: Array to store the derivatives of u.

u: Current position array of all spring boundary nodes.

p: Parameters tuple (N, kₛ, η, kf, l₀, δt, growth_dir).

t: Current time.

Description

Calculates the mechanical relaxation and normal velocity in a 1D system with periodic boundary conditions. The derivatives are based on spring forces and mechanical properties defined in p.

Equations for ODE problem:

TissueGrowth.δ — Method

δ(rᵢ₊₁, rᵢ)

Calculate the Euclidean distance between two points rᵢ₊₁ and rᵢ.

Arguments

rᵢ₊₁: The first point in space.

rᵢ: The second point in space.

Returns

The Euclidean distance between the two points.

TissueGrowth.ρ — Method

ρ(rᵢ₊₁, rᵢ)

Calculate the reciprocal of the distance (interpreted as density) between two points rᵢ₊₁ and rᵢ.

Arguments

rᵢ₊₁: The first point in space.

rᵢ: The second point in space.

Returns

The reciprocal of the distance between the two points.

TissueGrowth.τ — Method

τ(rᵢ₊₁, rᵢ₋₁)

Calculate the unit tangent vector between two neighboring points rᵢ₊₁ and rᵢ₋₁.

Arguments

rᵢ₊₁: The point after the central point in space.

rᵢ₋₁: The point before the central point in space.

Returns

The unit tangent vector between the two points.

TissueGrowth.n — Method

n(rᵢ₊₁, rᵢ₋₁, type)

Calculate the unit normal vector at a point rᵢ between two neighboring points rᵢ₊₁ and rᵢ₋₁. The orientation of the normal vector depends on the specified type.

Arguments

rᵢ₊₁: The point after the central point in space.

rᵢ₋₁: The point before the central point in space.

type: A string specifying the orientation of the normal vector, either "inward" or any other value for outward orientation.

Returns

The unit normal vector at the point.

TissueGrowth.Fₛ⁺ — Method

Fₛ⁺(rᵢ, rᵢ₊₁, rᵢ₋₁, kₛ, l₀)

Calculate the spring force (Nonlinear) for mechanical relaxation in the positive direction.

Arguments

rᵢ: The current point in space.

rᵢ₊₁: The point after the current point in space.

rᵢ₋₁: The point before the current point in space.

kₛ: Spring coefficient.

l₀: Resting length of the spring.

Returns

The spring force in the positive direction.

TissueGrowth.Fₛ⁻ — Method

Fₛ⁻(rᵢ, rᵢ₊₁, rᵢ₋₁, kₛ, l₀)

Calculate the spring force (Nonlinear) for mechanical relaxation in the negative direction.

Arguments

rᵢ: The current point in space.

rᵢ₊₁: The point after the current point in space.

rᵢ₋₁: The point before the current point in space.

kₛ: Spring coefficient.

l₀: Resting length of the spring.

Returns

The spring force in the negative direction.

TissueGrowth.Vₙ — Method

Vₙ(rᵢ₋₁, rᵢ, rᵢ₊₁, kf, δt, type)

Calculate the normal velocity of the interface such that Vₙ is proportional to ρ. The direction of the normal vector is determined by type.

Arguments

rᵢ₋₁: The point before the current point in space.

rᵢ: The current point in space.

rᵢ₊₁: The point after the current point in space.

kf: The amount of tissue produced per unit area per unit time.

δt: The time step.

type: A string specifying the orientation of the normal vector, either "inward" or any other value for outward orientation.

Returns

The normal velocity of the interface.

TissueGrowth.κ — Method

κ(rᵢ₋₁, rᵢ, rᵢ₊₁)

Approximate the curvature of a shape using the Menger method. This method is based on the areas of triangles formed by consecutive triplets of points.

Arguments

rᵢ₋₁: The point before the current point in space.

rᵢ: The current point in space.

rᵢ₊₁: The point after the current point in space.

Returns

The approximated curvature at the point rᵢ.

Reference

Anoshkina, Elena V., Alexander G. Belyaev, and Hans-Peter Seidel. "Asymtotic Analysis of Three-Point Approximations of Vertex Normals and Curvatures." VMV. 2002.

Cell behaviour functions:

TissueGrowth.P — Method

P(event, ρ, α)

Calculate the probability of proliferation occurring, given the density ρ and the parameter α.

If the proliferation event is considered to occur (event is true), the probability is calculated as the product of ρ and α. Otherwise, a vector of zeros is returned.

Arguments

event: A boolean indicating whether the proliferation event is considered to occur.

ρ: A vector representing densities.

α: Parameter for scaling the proliferation probability.

Returns

A vector representing the calculated probabilities for proliferation.

TissueGrowth.A — Method

A(event, ρ, β)

Calculate the probability of apoptosis (cell death) occurring, given the density ρ and the parameter β.

If the apoptosis event is considered to occur (event is true), the probability is calculated as the product of ρ and β. Otherwise, a vector of zeros is returned.

Arguments

event: A boolean indicating whether the apoptosis event is considered to occur.

ρ: A vector representing densities.

β: Parameter for scaling the apoptosis probability.

Returns

A vector representing the calculated probabilities for apoptosis.

TissueGrowth.E — Method

E(event, ρ, γ)

Calculate the probability of embedding occurring, given the density ρ and the parameter γ.

If the embedding event is considered to occur (event is true), the probability is calculated as the product of ρ and γ. Otherwise, a vector of zeros is returned.

Arguments

event: A boolean indicating whether the embedding event is considered to occur.

ρ: A vector representing densities.

γ: Parameter for scaling the embedding probability.

Returns

A vector representing the calculated probabilities for embedding.

TissueGrowth.affect! — Method

affect!(integrator)

Update the state of integrator based on probabilistic cellular events.

This function modifies integrator in place. It uses the parameters and state from integrator to compute probabilities for different cellular events: proliferation (prolif), death (death), and embedding (embed). Based on these probabilities and random draws, it updates the state of integrator.

Arguments

integrator: The integrator object containing the current state and parameters. The parameters are expected to be a tuple containing:

Returns

nothing. The function modifies integrator in place.

Custom modifier functions

Base.insert! — Function

Base.insert!(A::ElasticMatrix, col, xy)

Insert a two-element vector xy into a specific column col of an ElasticMatrix A.

This function modifies the ElasticMatrix in place. It inserts the elements of xy into the underlying data of A at the positions corresponding to the specified column.

Arguments

A::ElasticMatrix: The ElasticMatrix object to be modified.

col: The column index where the elements will be inserted.

xy: A two-element vector containing the values to be inserted into the matrix.

Returns

The modified ElasticMatrix A.

Base.deleteat! — Function

Base.deleteat!(A::ElasticMatrix, col)

Delete elements from a specific column col of an ElasticMatrix A.

This function modifies the ElasticMatrix in place. It removes the elements of the specified column from the underlying data of A.

Arguments

A::ElasticMatrix: The ElasticMatrix object to be modified.

col: The column index from which the elements will be removed.

Returns

The modified ElasticMatrix A.